# doc/api.yaml
openapi: "3.1.0"
info:
  title: Misinformation Backend API
  description: Check your Video, Image and Text authenticity
  version: "1.0.0"
  contact:
    name: Abegà Razafindratelo
    email: a.razafindratelo@gmail.com
    url: https://github.com/Abega1642
servers:
  - url: https://preprod-misinformation-by-ai-1.onrender.com
    description: Preprod server
  - url: https://misinformation-by-ai.onrender.com
    description: Prod server
...
paths:
...
  /api/media/videos/upload:
    post:
      operationId: uploadVideo
      tags:
        - Video
      summary: Upload a video file
      description: >
        Uploads a video file to the storage bucket.
        **Requires client credentials and Bearer token authentication.**
      security:
        - BearerAuth: [ ]
      parameters:
        - name: userEmail
          in: query
          required: true
          description: Email of the user uploading the video
          schema:
            type: string
            format: email
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Video file to upload
              required:
                - file
      responses:
        "200":
          description: Video uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "400":
          description: Invalid request - missing file or invalid email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: "2025-11-04T10:30:00Z"
                status: 400
                error: "Bad Request"
                message: "Video file is required and userEmail must be valid"
                path: "/api/media/videos/upload"
                errorCode: "BAD_FORM"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/InvalidClientCredentials"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/media/images/upload:
    post:
      operationId: uploadImage
      tags:
        - Image
      summary: Upload an image file
      description: >
        Uploads an image file to the storage bucket.
        **Bearer token authentication.**
      security:
        - BearerAuth: [ ]
      parameters:
        - name: userEmail
          in: query
          required: true
          description: Email of the user uploading the image
          schema:
            $ref: "#/components/schemas/Email"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file to upload
              required:
                - file
      responses:
        "200":
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Image"
        "400":
          description: Invalid request - missing file or invalid email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: "2025-11-04T10:30:00Z"
                status: 400
                error: "Bad Request"
                message: "Video file is required and userEmail must be valid"
                path: "/api/media/images/upload"
                errorCode: "BAD_FORM"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/InvalidClientCredentials"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  responses:
    Unauthorized:
      description: Authentication required - missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            timestamp: "2025-11-04T10:30:00Z"
            status: 401
            error: "Unauthorized"
            message: "Authentication required"
            path: "/api/endpoint"
            errorCode: "MISSING_AUTHORIZATION"
...
    InvalidClientCredentials:
      description: Invalid or missing client credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            timestamp: "2025-11-04T10:30:00Z"
            status: 403
            error: "Forbidden"
            message: "Invalid or missing client credentials"
            path: "/api/endpoint"
            errorCode: "AUTHORIZATION_DENIED"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            timestamp: "2025-11-04T10:30:00Z"
            status: 500
            error: "Internal Server Error"
            message: "An unexpected error occurred"
            path: "/api/endpoint"
            errorCode: "INTERNAL_SERVER_ERROR"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Include in Authorization header as: `Bearer <token>`
  schemas:
    SizeType:
      type: string
      enum:
        - BYTES
        - KB
        - MB
        - GB
        - TB
        - PB
        - EB
    AudioCodec:
      type: string
      enum:
        - NONE
        - AAC
        - MP3
        - MP2
        - AC3
        - EAC3
        - DTS
        - PCM
        - PCM_S16LE
        - PCM_S24LE
        - FLAC
        - OPUS
        - VORBIS
        - WMA
        - ALAC
        - AMR_NB
        - AMR_WB
        - G711
        - G722
        - UNKNOWN
    ContainerFormat:
      type: string
      enum:
        - MP4
        - MKV
        - MOV
        - AVI
        - FLV
        - WMV
        - WEBM
        - MPEG_TS
        - MPEG_PS
        - THREEGP
        - OGG
        - M4A
        - WAV
        - FLAC
        - ASF
        - APE
        - AIFF
        - UNKNOWN
    Codec:
      type: string
      enum:
        - H264
        - H265
        - MPEG1
        - MPEG2
        - MPEG4
        - VP8
        - VP9
        - AV1
        - THEORA
        - WMV1
        - WMV2
        - WMV3
        - MJPEG
        - PRORES
        - DNXHD
        - CINEFORM
        - DV
        - XVID
        - DIVX
        - HEVC
        - UNKNOWN
    FileType:
      type: string
      enum:
        - VIDEO
        - IMAGE
        - AUDIO
    UUID:
      type: string
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      format: uuid
      example: e6b5cc3e-eaaa-4cb3-912c-e059e983bb78
    Date:
      type: string
      format: date-time
      description: 'Date and time in ISO 8601 format (RFC 3339)'
      example: '2025-12-06T11:23:45Z'
    Email:
      type: string
      format: email
      pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$'
      example: a.razafindratelo@gmail.com
      description: User's email address
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          $ref: "#/components/schemas/Date"
          description: When the error occurred
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: HTTP status reason phrase
          example: "Bad Request"
        message:
          type: string
          description: Detailed error message
          example: "Validation failed"
        path:
          type: string
          description: Request path where error occurred
          example: "/path/of/the/endpoint"
        errorCode:
          type: string
          description: Application-specific error code
          enum:
            - AUTHORIZATION_DENIED
            - ENTITY_NOT_FOUND
            - CONSTRAINT_VIOLATION_ON_FIELDS
            - TOKEN_UNIQUENESS_CONFLICT
            - BAD_FORM
            - MISSING_AUTHORIZATION
            - INVALID_AUTHORIZATION_FORMAT
            - INVALID_TOKEN
            - TOKEN_NOT_FOUND
            - USER_DEACTIVATED
            - AUTHENTICATION_FAILED
            - INTERNAL_SERVER_ERROR
            - MISSING_REQUIRED_PARAMETER
            - INVALID_PARAMETER
            - MALFORMED_JSON
            - METHOD_NOT_ALLOWED
      required:
        - timestamp
        - status
        - error
        - message
        - path
...
    User:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        email:
          $ref: "#/components/schemas/Email"
        clerk_id:
          $ref: "#/components/schemas/UUID"
        creation_date:
          $ref: "#/components/schemas/Date"
...
    ImageFormat:
      type: string
      enum: [ JPEG, PNG, GIF, BMP, TIFF, WEBP, RAW, SVG ]
    Image:
      allOf:
        - $ref: "#/components/schemas/Media"
        - type: object
          required:
            - width
            - height
            - format
            - bitDepth
            - colorModel
          properties:
            width:
              type: integer
              format: int32
              description: Image width in pixels
            height:
              type: integer
              format: int32
              description: Image height in pixels
            format:
              $ref: "#/components/schemas/ImageFormat"
            bitDepth:
              type: integer
              format: int32
              description: Color depth in bits per pixel
            colorModel:
              type: string
              enum: [ RGB, CMYK, Grayscale, LAB, HSV, Indexed ]
              description: Color model/space used
            dpi:
              type: number
              format: double
              description: Dots per inch (resolution)
            iso:
              type: integer
              format: int32
              description: ISO sensitivity (for photos)
    Media:
      type: object
      discriminator:
        propertyName: fileType
      required:
        - id
        - fileName
        - size
        - sizeType
        - fileType
        - createdAt
        - bucketKey
        - owner
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        fileName:
          type: string
          description: Name of the media file
        size:
          type: number
          format: double
          description: Size of the file in bytes
        sizeType:
          $ref: "#/components/schemas/SizeType"
        fileType:
          $ref: "#/components/schemas/FileType"
        createdAt:
          $ref: "#/components/schemas/Date"
        bucketKey:
          type: string
          description: Path to the file in storage
        owner:
          $ref: "#/components/schemas/User"
    Video:
      allOf:
        - $ref: "#/components/schemas/Media"
        - type: object
          required:
            - duration
            - codec
            - width
            - height
            - containerFormat
          properties:
            duration:
              type: number
              format: double
              description: Video duration in seconds
            codec:
              $ref: "#/components/schemas/Codec"
            width:
              type: integer
              format: int32
              description: Video width in pixels
            height:
              type: integer
              format: int32
              description: Video height in pixels
            frameRate:
              type: number
              format: double
              description: Frames per second
            aspectRatio:
              type: string
              description: Display aspect ratio (e.g., "16:9", "4:3")
            containerFormat:
              $ref: "#/components/schemas/ContainerFormat"
            bitRate:
              type: number
              format: double
              description: Bit rate in bits per second
            audioCodec:
              $ref: "#/components/schemas/AudioCodec"
            audioChannels:
              type: integer
              format: int32
              description: Number of audio channels
            audioSampleRate:
              type: integer
              format: int32
              description: Audio sample rate in Hz
...
security:
  - BearerAuth: [ ]